requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
requests>=2.31.0
pydantic>=2.4.0
python-multipart>=0.0.6
pandas>=2.1.0
google-genai
langsmith
python-dotenv>=1.0.0
discord.py
